backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
Pillow>=10.0
google-auth[requests]>=2.0.0
